Use watson for micro issues
`watson` is a tiny inline issue manager, which can grep through source
and report things that need to be done. Its basically a pretty looking
grep.

There is no need for everyone to use it, but its good to have. The only
extra effort is to annotate all FIX/TODO/REVIEW in a consistent manner,
which will help even if you don't use the tool. See the diff for
details.

See the details here: http://goosecode.com/watson/

As of now, a sample output from watson looks like this:

    ------------------------------
    watson - inline issue manager
    Run in: /Users/j/Projects/functorrent
    Run @ Sun Nov 15 01:07:06 2015
    ------------------------------

    [ x ] src/FuncTorrent/Bencode.hs
    [ todo ]
      line 156 - Use builders and lazy byte string to get O(1) concatenation over O(n)
      line 159 - encode . decode pair might be a good candidate for Quickcheck.

    [ x ] src/FuncTorrent/ControlThread.hs
    [ todo ]
      line 81 - This will cause a 4s delay b/w a ^C and the app going down
      line 121 - Write to ~/.functorrent/caches

    [ x ] src/FuncTorrent/Network.hs
    [ todo ]
      line 16 - Url encode each argument

    [ x ] src/FuncTorrent/Tracker.hs
    [ review ]
      line 74 - Write urlEncodeHash with Network.HTTP.Base.urlEncode or something

    [ todo ]
      line 73 - Add tests for URL encode hash

    [ x ] test/PeerThreadMock.hs
    [ todo ]
      line 17 - Non exhaustive pattern match